kunzite/requirements.txt
ansible
pyyaml
pip
wheel
